Adding version upper bounds in install_requires
Our version upper bound on taurus-citrine was removed
with the last release, making any upstream breaking changes breaking
to this package as well.  This hotpatch adds the upper version bound
back (on the minor rather than major version since taurus-citrine
is pre-release) and also adds very generous upper bounds to the
rest of the dependencies.

py.test was removed as a install_requires because it is actually
a tests_require.  An issue has been opened to add those dependencies: #181