Not assuming UID == a unique idenitifer for an entity
Listing the versions of a display table will return an entity that has
multiple entities with the same UID, but different versions. Expand the
notion of equality based on a subset of the entity we extract (for
comparison), and stop iterating if we find a cycle.

We should definitely log when this happens, as it's an internal error.