Fix ProjectCollection.search; use in find_collection
This adds a `get_resource_post` method to the session that behaves
like get_resource (in that it extracts the response json) but
POSTs with a body instead of GET.  That method is then
used in ProjectCollection.search.

The find_collection method will now use the search method when
the collection is a ProjectCollection, since that is more efficient.

Most of the changes are to maintain test coverage.