Change the design space size check to count cells
The client-side design space size limit is meant simply to catch
very large POSTs and PUTs that would cause a 5xx on the backend
because they are too big.  This changes "too big" from either
"too many rows" or "too many columns" to "too many cells", with
the actual number of cells unchanged.

NOTE this does not imply that we can actually handle any
EnumeratedDesignSpace that fits within this limit; simply that it
shoudln't cause a 5xx in the handler on the backend.