Deduplicate list results by their uid
Since we're doing pagination-based listing, creates and deletes
to the collection being listed can result in index movement during
a multi-request list operation.  In particular, the repetition of
a list element from the end of one page to the beginning of the next
has been observed to occur when items are added to the list.  This
change keeps a list of the uids that have been encountered and skips
returning duplicates. If the listed object isn't a resource and
therefore doesn't have a uid, then no deduplication is performed.

This behavior can be toggled via a flag to assist future developers.