Use str() instead of __class__ to avoid constructing resources
The existing logic tried to construct a resource with an empty
constructor and then use __class__ to figure out the class name.
This failed when the constructor had required arguments.

The change is to just use str() on the class, which should have
the same behavior and avoid hitting the constructor.