Deprecate Score's name and description for future removal
The name and description fields on scores are confusing since
they are not modules (i.e. assets or workflows).  We're removing
them, but want to do so with a deprecation cycle.  This deprecates
setting and getting them.

Note that there is a breaking change here: because name and
description were the first two arguments in the constructor,
they can't simply be given defaults in the first two positions.
To resolve this, the PR adds the keyword-only setting to the
constructors, which is a best practice anyways.