Move _session and _project_id to base class; report lazy
This removes the session and report from predictor constructors
and puts them in the predictor base class.  When a predictor
is gotten, the project_id and session from the collection is inserted
so that subsequent calls to report can lazily load the report contents.

Tests were upated.  Since this is in the base class, a lot
of test content was able to be removed.