[PLA-8726] Adding support for branches.
This includes updating design workflows to use the v2 endpoints and
report branch_id wherever possible. The one exception is creating a
design workflow, where we must hit the v1 endpoint if the user hasn't
indicated a branch and then retrieve that branch_id after the fact.

Note that archiving branches is not supported by this commit; that will
be added in a follow on.