Mark design_space_id and predictor_id optional.
The v2 design workflow API supports "partial workflows", which is one
which isn't valid, but allows the user to save their state. To support
this, null values for design_space_id and predictor_id are allowed on
creation, update, and retrieval. As such, we allow DesignWorkspace
objects to have those values (as well as processor_id) set to None.