[PLA-9127] AutoMLPredictor should support all outputs.
The AutoML predictor payload contains a list of outputs, all of which
should be available to the user. This involves deprecating the output
field, which only contains the first element of the list, in favor of a
new field containing the whole list.

Also removed dead code dealing with the responses field of the payload.