[PLA-9117] Remove the experimental fields.
It's been hard-coded in the backend to always be false with an empty
list of reasons. It's not accessed in many places, but we do need to
deprecate it where we do. But mostly, this just rips it out and updates
the tests, including adding tests that a deprecation warning is raised.