[PLA-9375] Fix create_or_update for design workflow.
seeding.find_or_create.create_or_update attempts to retrieve the passed
resource by name. If successful, the original object is copied, and the
uid is copied from the retrieved resource before updating. We also need
to copy the branch_id if it's a DesignWorkflow, since the branch ID is
required for performing an update.