Update predictor to properly present archive status.
In the near future, we'll be adding support for predictor versions.
There will be a separation between a predictor version and the predictor
root, which can be archived independently. However, the predictor root
will be invisible to the user. Since this logic already exists in the
backend, that means the value of "is_archived" may no longer be the
expected value.

Until we expose that separation and make this explicit, we should present
the root's archive status, which matches what the user previously
expected. In addition, we point users at the new methods, which act on
the predictor root.