[PLA-12391] Support retraining stale predictors.
Our releases sometimes mean the platform cannot load trained predictors.
We call these predictors stale. The predictor configuration is still in
tact and usable, so our solution is to silently retrain it when it's
used in a design execution.

The modify candidates feature allows users to use an old, trained
predictor directly. As such, it also sometimes requires retraining, but
outside the context of a design execution. As such, this exposes a route
to check if a predictor is stale, and to retrain such a predictor.