Deprecate dehydrated design spaces.
Dehydration is a reuse strategy which we don't really support any more.
In fact, using it will result in an error when registering or updating a
design space. However, since it's still documented, we should document
it and do the right thing for now, but drop it in the next major
release.