[PLA-12982] Update tests.
Add new ones, update existing ones to use the new interfaces, and mark
the old interfaces and their uses as deprecated.

Also pin pint. That library is still in beta, and the way gemd currently
references it grabs the latest version, leading us pulling in breaking
changes. And we can't upgrade gemd just yet, as we still support Python
3.7 and gemd doesn't.