[PLA-13149] Bump dependency versions.
Bumped all dependencies to their latest version, with a few exceptions.

pytest-flake8 and pydocstyle don't appear to be in use, so they were
dropped instead.

botocore only needs to be in our requirements file if we need to pin it
separately from boto3, which we currently do not.

boto3 imposes some restrictions on urllib3. If Python is < 3.10, urllib3
must be pre-2.0. Since 1.31.62, if Python is more recent, it can support
up to (but not including) urllib3 2.1. We've set the pins accordingly.

The latest version of Sphinx available for Python 3.8 is 7.1.2, so we
pin it to that version across the board.