[PLA-13752] Add link for API Reference.
Since we're generating our API docs, one of our tables of contents must
link to it, or else it will throw a warning. So I've added it to the
main page. That main page TOC goes multiple levels deep, so it will show
the generated page's name. Since we were starting it at src/, that was
its name, which wouldn't make much sense to users. So I tweaked it to
start at src/citrine/.