[PLA-13814] Switch from Travis to GitHub Actions.
We've had some issues with Travis CI, most recently permissions syncing.
In the past few years, GitHub Actions has become pretty mature, and it
has documentation that's generally far better than Travis. Not to
mention its more tightly integrated with GitHub for obvious reasons. And
the general benefits of shrinking the number of services we depend on.

This introduces three workflows, which will be kicked off automatically.
One runs the tests and linting and such on PR creation against main, one
builds and deploys to PyPI when a release is created, and one builds and
deploys the docs when a release is created.