Add safety check for updated dependencies.
As an extra safety check, in the Travis CI config, we would also do a
run which upgraded some of our dependencies to their latest version,
then ran the tests.

This does the same, except it installs from the setup.py to ensure we're
only testing the max version we claim to support, which usually means
the latest minor version up to just before some major version. Also, the
jobs are marked as continue-on-error, so they should show up as failing
but not block merging.