Surface docs warnings.
In order to keep the docs build messages useful, we should run a check
on PR to surface all warnings. With the release of 3.0, they should be
completely clean, so a failure of this job will be meaningful. Even
such, we'll probably keep it non-required for now so as to not add
unecessarily to the development burden.