Tighten the type discriminators.
Some of them are looser than necessary. This won't constitute a breaking
change, as the resulting field type will be the same, just without using
the polymorphic decoding.

The idea is to increase precision, as well as allow us to drop
unnecessary type discriminators from the payload in the future.