[PNE-241] Data Manager support.
Datasets and all related objects now belong to teams, so new endpoints
have been introduced into the platform to support this. As such, all
such calls need to be updated. The result is a bunch of
`ProjectCollection` methods are being deprecated in favor of their
`TeamCollection` equivalents, and many methods which accepted a project
or project_id must be moved to a team/team_id.

The vast majority of behaviors will remain unchanged when using the
deprecated methods, until we bump the major version of the SDK and drop
the old endpoints. The major exception is that creating a dataset on a
project endpoint will register it to a team, so it will be omitted when
you list your datasets by project.