Deprecate training_data on subpredictors.
For quite a while, the backend has only supported training data supplied
on graph predictors. To support the existing interface where it can be
provided on many subpredictors, the backend hoists any training data it
finds up to the root.

Deprecating the associated fields in the SDK is a step towards no longer
needing to do that.