Adjust type of ChemicalFormulaFeaturizer.powers.
The backend is opening up the type of powers to be floats. As such, this
commit will deserialize powers as a list of floats. However, to preserve
backwards compatability, the type of powers will be preserved, and a
new field (powers_float) with the list as floats is introduced. In
v4.0.0, the type of 'powers' will be changed, and 'powers_float' will be
deprecated before its removal in v5.0.0.