[PNE-6647] Tag default DS with settings object.
When creating a design space, the endpoint accepts a dict of the
settings which were used to create it. To make this process transparent
from a user perspective, any design space which is created with
create_default will have the settings dict attached to it in non-public
field. So when the user goes to register it, the settings object will
automatically be included.

Additionally, when retrieving a design space from the backend, those
settings are part of the metadata. They will be captured in a non-public
field, so if a user turns around and registers the design space object
again, this copy will contain those settings.