[PNE-7018] Add new evaluations API.
PEWs no longer need to be a customer facing asset type. Their only
function is to allow reusing the same set of evaluators with different
predictors, which A) doesn't often work due to mismatched responses,
and B) isn't particularly helpful. As such, we can move users on to a
cleaner, easier to use workflow, where predictor evaluation is kicked
off directly against predictor, and you can start one using the default
evaluators with a single call.

The PredictorEvaluation API and associated objects are similar to the
existing PEW and PEW execution APIs/objects, but are stand-alone. This
ensures we can easily make a clean break when the deprecation cycle
completes, and also provides an opportunity to ensure the interfaces are
as clean and simple as possible.