[PNE-7661] Deprecate most top-level design spaces.
As a simplification for the platform, all design spaces should be either
a ProductDesignSpace or HierarchicalDesignSpace. The other types
(namely, DataSourceDesignSpace and FormulationDesignSpace) should be
subspaces of them. This is the only structure supported by the UI, so it
will improve their experience of the platform, too.

Additionally, EnumeratedDesignSpace is obsolete, and should be replaced
by a DataSourceDesign space within a ProductDesignSpace.

Note that we also emit a warning on retrieval. That's because when we go
to v4.0, we will stop supporting such design spaces at all. This gives
users a chance to update their design spaces as necessary, and ensures
they will not be surprised when the type of design space they retrieve
changes.